Create Partition Labels
Problem statement
You are given a string ‘S’ of lowercase English letters. Your task is to partition this string into as many parts as possible so that all the occurrences of the same letter appear in exactly one of the partitions.

For example:
"qvmwtmzzse" is the given string. We can make the first two partitions as ‘q’, ‘v’, but ‘m’ can not be an individual partition because there is another occurrence of m  at index 6. So, the minimum length of the third partition is from index 3 to 6 and the partition is "mwtm". Similarly, the next partitions are "zz", ‘s’, ‘e’. 
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1<= T <= 10^3
1 <= |S| <= 10^3

Where |S| is the size of the given string ‘S’.

Time Limit: 1 sec
Sample Input 1:
2
qvmwtmzzse
dccccbaabe
Sample Output 1:
1 1 4 2 1 1
1 4 4 1 
Explanation Of Sample Input 1:
In the first test case, the given string is “qvmwtmzzse”. To maximize the number of partitions, we will make partitions as follows “q”, “v”, “mwtm”, “zz”, “s”, “e” and the lengths are 1, 1, 4, 2, 1, 1 respectively. Hence the answer is {1, 1, 4, 2, 1, 1}.

In the second test case, the given string is “dccccbaabe”. To maximize the number of partitions, we will make partitions as follows “d”, “cccc”, “baab”, “e” and the lengths are 1, 4, 4, 1 respectively. Hence the answer is {1, 4, 4, 1}.
Sample Input 2:
2
vhaagbqkaq
eaaaabaaec
Sample Output 2:
1 1 8
9 1
Explanation Of Sample Input 2:
In the first test case, the given string is “vhaagbqkaq”. To maximize the number of partitions, we will make partitions as follows “v”, “h”, “aagbqkaq” and the lengths are 1, 1, 8 respectively. Hence the answer is {1, 1, 8}.

In the second test case, the given string is “eaaaabaaec”. To maximize the number of partitions, we will make partitions as follows “eaaaabaae”, “c” and the lengths are 9, 1 respectively. Hence the answer is {9, 1}.