Create  Tic Tac Toe Winner
Problem statement
Ninja is bored and hence started playing tic tac toe with his friend. As always he started losing the game. So he approaches you to find the optimal move he needs to play at a given moment so that he can defeat his friend.

You are given the current state of the game with a 3x3 grid, of which some squares are filled with 'X', some are filled with ‘O’ and the rest are filled with ‘_’ which means that this square is empty. You need to find the best possible score our ninja can get and the position of the optimal move he should make. You always start first and your symbol is ‘X’. The scores are assigned as follows:

Win: 1

Draw : 0

Lose : -1

For Example:
Suppose the board looks as follows: 

X O X
O O X
_ _ _


We can see that if we put X at positions 3, 3 in the grid we win straightaway. Hence the best possible score is 1 (Win). If we put our symbol anywhere else we will not be able to achieve the desired score.
Note :
If there are multiple positions leading to the same score, choose the position with the minimum row. If there are still multiple positions, choose the one with the minimum column.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
The size of the grid will always be 3.

Time Limit: 1 sec.
Sample Input 1:
X O X
O O X
_ _ _
Sample Output 1:
1 3 3
Explanation For Sample Output 1:
The value of the best move is 1 (i.e win). This can be achieved by making a move at position 3, 3. This way the column 3 will contain all X and we will get the required score. 
Sample Input 2:
O _ O 
_ X _
O X X
Sample Output 2:
-1 1 2