Create Maximum Score
Problem statement
Ninja is playing a board game in which two lists of distinct numbers ‘A’ and ‘B’ arranged in a non-descending order are given. The game has certain rules and the player has to pick some numbers from the given list and the score is the sum of unique picked numbers. The rules are:

1.  Choose any list ‘A’ or ‘B’.
2.  Traverse from left to right.
3.  After picking a number, if the picked number is present in both the arrays, you are allowed to traverse to the other array.
You are given arrays,’A’ and ‘B’ of size ‘N’ and ‘M’ respectively. Your task is to find the maximum score Ninja can achieve.

Since the answer can be very large, print answer % (10^9 + 7).
For Example
If the given array are ‘A’ = [1,3,5,7,9] and  ‘B’ = [3,5,100]”.The maximum score can be achieved is 109[1+3+5+100].
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 10
1 <= N,M <= 10^4.
1 <= A[i], B[i] <= 10^6.  

Time limit: 1 sec
Sample Input 1:
2
5 3
1 3 5 7 9
3 5 100
4 3
1 2 3 5
1 3 10 
Sample Output 1:
109
16
Explanation of sample input 1:
For the first test case,

The numbers Ninja will be pick are [1,3,5,100].The sum of unique numbers are 109.
Hence, the answer is 109. 

For the second test case:

 The numbers Ninja will be pick are [1,2,3,3,10].The sum of unique numbers are 16(1+2+3+10).
 Hence, the answer is 16. 
Sample Input 2:
2
5 4
2 4 5 8 10
4 6 8 9
5 5
1 2 3 4 5
6 7 8 9 10
Sample Output 2:
30
40