Create Check if Possible to survive
Problem statement
You are a poor citizen of an island that has only one shop for groceries, which is open six days a week and is closed on Sundays. The first day of the week is Monday. You can buy ‘n’ units of food from a day’s work. Now each day, you need ‘x’ units of food to survive. You are making a boat to escape the island, which will require ‘d’ days to be made. You need to survive at least ‘d’ days on the island. You are given three integers n, x, and d. Your task is to determine if you can escape the island

Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 10^4
1<= n, m, d<=10^8

Where ‘T’ is the total number of test cases and n denotes the units of food the citizen can buy from one day salary, x denotes the units needed to survive each day and d denotes the minimum number of days the citizen needs to survive on the island to escape it.   

Time limit: 1 sec
Sample Input 1:
2
16 2 10
2 4 20
Sample Output 1:
YES
NO
Explanation of sample input 1 :
Test Case 1:

One possible solution is to buy food on the first day (Monday). It’s sufficient to eat from this food up to the 8th day (Monday) inclusive. On the 9th day (Tuesday), repurchase food and use the food to survive the 9th and 10th day.

Test Case 2:

We can not survive as we make only two units of food per day, but we consume four units of food, so we will not be able to survive.
Sample Input 2:
2
10 2 30
10 9 5
Sample Output 2:
YES
YES