Create Save The World
Problem statement
We are in the endgame now. Tony and Thanos are fighting. Before the war, Captain Marvel secretly gave Tony a superpower that she took from the other universe. They went through some days of the war but Tony’s power was enough to fight against Thanos but not to defeat him and vice versa. So, they decided to play a number game.

In the game, there are two integers - ‘maximumNumber’, and ‘target’. In each turn, one player can choose an integer and add it to ‘runningTotal’. The player who makes the ‘runningTotal’ equal to or more than the ‘target’ wins.

The rules of the game are:

1. One can use any number between 1 and ‘maximumNumber’.
2. One integer can’t be used more than once.
3. The running total = 0 initially.
4. Both players play optimally.
Thanos offered Tony to make the first move and he will choose the two integers. But the avengers are not satisfied. So, all came to a conclusion that after choosing the integers if Tony is not satisfied with the two integers, he will tell Thanos to rechoose two different integers. Tony is completely exhausted but doesn’t want to lose as it can cost us the complete universe. Help Tony to figure out if he can win or not from the two integers that Thanos chose.

Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 50
1 <= maximumNumber <= 20
1 <= target <= 200

Time Limit: 1 sec
Sample Input 1:
2
5 2
5 6
Sample Output 1:
1
0
Explanation for Sample Input 1:
In the first test case, It’s Tony’s turn, he can choose any number between 1 and 5. Initially runningTotal = 0. He will choose any number between 2 and 5 and win the game. So, the output is 1.

In the second test case, Initially runningTotal = 0. It’s Tony’s turn, he has to choose any number between 1 and 5. Whichever number he will choose runningTotal will lie in the range 1 to 5. In the next turn, Thanos will choose a number between 1 and 5 which Tony didn’t choose in his turn, and win the game. So, the output is 0.
Sample Input 2:
2
10 15
10 22
Sample Output 2:
1
0