Create Numbers with product K
Problem statement
Ninja Ankush likes to brag that he is the Ultimate Ninja among his peers. Therefore his fellow Ninja Nikhil gave him a riddle to check if Ankush is really the Ultimate Ninja. Nikhil gave Ankush a range and a number ‘K’, and asked how many numbers exist in the range such that the product of the digits of the number is equal to ‘K’. Help Ninja Ankush to prove to Ninja Nikhil that he, in fact, is the Ultimate Ninja.

More Formally, Given three positive integers ‘L’, ‘R’ and ‘K’, the task is to count the numbers in the range ‘L’ and ‘R’ inclusive, whose product of digits is equal to ‘K’.

For example

Given:
‘L’ = 1, ‘R’ = 23, ‘K’ = 6.

The answer will be 3 since there are three numbers between 1 and 23 whose product of digits is 6, and those are 6, 16, and 23.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= ‘T’ <= 10
1 <= ‘L’ <= 10 ^ 8
‘L’ <= ‘R’ <= 10 ^ 8
1 <= ‘K’ <= 10 ^ 4

Time Limit: 1sec.
Sample Input 1 :
2
1 23 6
11 25 2
Sample Output 1 :
3
2 
Explanation of the Sample Input 1:
In the first test case, The answer will be 3 since there are 3 numbers between 1 and 23 whose product of digits is 6, and those are 6, 16, and 23.

In the second test case, The answer will be 2 since there are 2 numbers between 1 and 25 whose product of digits is 2, and those are 12, 21.
Sample Input 2 :
2
3 23 3
4 50 9
Sample Output 2 :
2
3