Create Count Of 3s
Problem statement
You are given an integer ‘N’. You simply need to find out the number of occurrences of 3 as a digit in the range of numbers from [0, N].

Note:
 You need to count occurrences at every place of the number.
For example :
You are given N = 13, then the number of occurrences of 3 in range [0, 13] = 2 (3, 13), you need to return 2.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 10^2
0 <= N <= 10^5

where ‘T’ is the number of test cases and 'N' is the given integer.

Time Limit: 1 sec
Sample Input 1:
2
13
24
Sample Output 1:
2
3
Explanation For Sample Input 1:
In the first test case, 
Number of occurrences of 3 in range [0, 13] = 2 (3, 13). Return 2

In the second test case, 
Number of occurrences of 3 in range [0, 24] = 3 (3, 13, 23). Return 3
Sample Input 2:
2
10
33
Sample Output 2:
1
8
Explanation For Sample Input 2:
In the first test case, 
Number of occurrences of 3 in range [0, 10] = 1 (3). Return 1

In the second test case, 
Number of occurrences of 3 in range [0, 33] = 8 (3, 13, 23, 30, 31, 32, 33). Return 8